requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cloudinary==1.44.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==44.0.2
dj-database-url==2.3.0
Django==5.1.3
django-allauth==65.6.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
et_xmlfile==2.0.0
gunicorn==23.0.0
idna==3.10
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysqlclient==2.2.5
openpyxl==3.1.5
packaging==24.2
pillow==11.0.0
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.9.0
python-decouple==3.8
requests==2.32.3
setuptools==75.8.2
six==1.17.0
sqlparse==0.5.1
typing_extensions==4.13.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
whitenoise==6.9.0
